requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
azure-identity==1.15.0
azure-cosmos==4.5.1
openai==1.3.7
pydantic==2.5.0
aiofiles==23.2.1
